split ResultAction from MatchAction
