+++ idaplugin/rematch/actions/match.py
@@ -145,81 +145,4 @@ class MatchAction(base.BoundFileAction):
 
     self.clean()
 
-    self.start_results()
-
-  def start_results(self):
-    self.pbar = QtWidgets.QProgressDialog()
-    self.pbar.setAutoReset(False)
-    self.pbar.canceled.connect(self.cancel)
-    self.pbar.rejected.connect(self.cancel)
-    self.pbar.setLabelText("Receiving match results...")
-    self.pbar.setRange(0, 3)  # 3 for 3 end points (locals, remotes, matches)
-    self.pbar.setValue(0)
-    self.pbar.accepted.connect(self.accept_results)
-    self.pbar.show()
-
-    self.results = MatchResultDialog(self.task_id, "Match results")
-
-    self.seen = set()
-
-    log('match_action').info("Result download started")
-    locals_url = "collab/tasks/{}/locals/".format(self.task_id)
-    q = network.QueryWorker("GET", locals_url, json=True, pageable=True,
-                            params={'limit': 100})
-    q.start(self.handle_locals)
-    self.delayed_queries.append(q)
-
-    remotes_url = "collab/tasks/{}/remotes/".format(self.task_id)
-    q = network.QueryWorker("GET", remotes_url, json=True, pageable=True,
-                            params={'limit': 100})
-    q.start(self.handle_remotes)
-    self.delayed_queries.append(q)
-
-    matches_url = "collab/tasks/{}/matches/".format(self.task_id)
-    q = network.QueryWorker("GET", matches_url, json=True, pageable=True,
-                            params={'limit': 200})
-    q.start(self.handle_matches)
-    self.delayed_queries.append(q)
-
-  def handle_locals(self, response):
-    new_locals = {obj['id']: obj for obj in response['results']}
-    self.results.add_locals(new_locals)
-
-    self.handle_page(response, 'locals')
-
-  def handle_remotes(self, response):
-    new_remotes = {obj['id']: obj for obj in response['results']}
-    self.results.add_remotes(new_remotes)
-
-    self.handle_page(response, 'remotes')
-
-  def handle_matches(self, response):
-    def rename(o):
-      o['local_id'] = o.pop('from_instance')
-      o['remote_id'] = o.pop('to_instance')
-      return o
-
-    new_matches = map(rename, response['results'])
-    self.results.add_matches(new_matches)
-
-    self.handle_page(response, 'matches')
-
-  def handle_page(self, response, page_type):
-    if page_type not in self.seen:
-      self.seen.add(page_type)
-      self.pbar.setMaximum(self.pbar.maximum() + response['count'])
-      self.pbar.setValue(self.pbar.value() + 1)
-
-    self.pbar.setValue(self.pbar.value() + len(response['results']))
-    log('match_action').info("result download progress: {} / {} with {}"
-                             "".format(self.pbar.value(), self.pbar.maximum(),
-                                       self.seen))
-    if self.pbar.value() >= self.pbar.maximum():
-      self.pbar.accept()
-
-  def accept_results(self):
-    # log('match_action').info("Result download completed successfully")
-
-    self.clean()
-
-    self.results.show()
+    ResultAction(self.task_id).activate()
