+++ idaplugin/rematch/dialogs/__init__.py
@@ -4,9 +4,10 @@ from . import gui
 from . import login
 from . import upload
 from . import match
+from . import result
 from . import project
 from . import settings
 from . import serializedgraph
 
-__all__ = ['base', 'widgets', 'gui', 'login', 'upload', 'match', 'project',
-           'settings', 'serializedgraph']
+__all__ = ['base', 'widgets', 'gui', 'login', 'upload', 'match', 'result',
+           'project', 'settings', 'serializedgraph']
